wip: rename items to tree
