+++ src/apps/fuse/dependency-injection/virtual-drive/tree/TreeContainer.ts
@@ -1,4 +1,4 @@
-import { TreeBuilder } from '../../../../../context/virtual-drive/items/application/TreeBuilder';
+import { TreeBuilder } from '../../../../../context/virtual-drive/tree/application/TreeBuilder';
 
 export interface TreeContainer {
   existingNodesTreeBuilder: TreeBuilder;
