+++ src/apps/hydration-api/dependency-injection/common/traverser.ts
@@ -1,5 +1,5 @@
 import crypt from '../../../../context/shared/infrastructure/crypt';
-import { Traverser } from '../../../../context/virtual-drive/items/application/Traverser';
+import { Traverser } from '../../../../context/virtual-drive/tree/application/Traverser';
 import { DependencyInjectionUserProvider } from './user';
 
 export class DependencyInjectionTraverserProvider {
