+++ src/apps/hydration-api/dependency-injection/tree/builder.ts
@@ -1,9 +1,9 @@
-import { Traverser } from '../../../../context/virtual-drive/items/application/Traverser';
-import { TreeBuilder } from '../../../../context/virtual-drive/items/application/TreeBuilder';
-import { CryptoJsNameDecrypt } from '../../../../context/virtual-drive/items/infrastructure/CryptoJsNameDecrypt';
+import { Traverser } from '../../../../context/virtual-drive/tree/application/Traverser';
+import { TreeBuilder } from '../../../../context/virtual-drive/tree/application/TreeBuilder';
+import { CryptoJsNameDecrypt } from '../../../../context/virtual-drive/tree/infrastructure/CryptoJsNameDecrypt';
 import { getUser } from '../../../main/auth/service';
 import { TreeContainer } from './TreeContainer';
-import { SQLiteRemoteItemsGenerator } from '../../../../context/virtual-drive/items/infrastructure/SQLiteRemoteItemsGenerator';
+import { SQLiteRemoteItemsGenerator } from '../../../../context/virtual-drive/tree/infrastructure/SQLiteRemoteItemsGenerator';
 
 export function buildTreeContainer(): TreeContainer {
   const user = getUser();
