+++ src/apps/sync-engine/dependency-injection/items/ItemsContainer.ts
@@ -1,4 +1,4 @@
-import { TreeBuilder } from '../../../../context/virtual-drive/items/application/TreeBuilder';
+import { TreeBuilder } from '../../../../context/virtual-drive/tree/application/TreeBuilder';
 
 export interface ItemsContainer {
   existingItemsTreeBuilder: TreeBuilder;
