+++ src/apps/sync-engine/dependency-injection/items/builder.ts
@@ -1,7 +1,7 @@
-import { IpcRemoteItemsGenerator } from '../../../../context/virtual-drive/items/infrastructure/IpcRemoteItemsGenerator';
-import { Traverser } from '../../../../context/virtual-drive/items/application/Traverser';
-import { TreeBuilder } from '../../../../context/virtual-drive/items/application/TreeBuilder';
-import { CryptoJsNameDecrypt } from '../../../../context/virtual-drive/items/infrastructure/CryptoJsNameDecrypt';
+import { IpcRemoteItemsGenerator } from '../../../../context/virtual-drive/tree/infrastructure/IpcRemoteItemsGenerator';
+import { Traverser } from '../../../../context/virtual-drive/tree/application/Traverser';
+import { TreeBuilder } from '../../../../context/virtual-drive/tree/application/TreeBuilder';
+import { CryptoJsNameDecrypt } from '../../../../context/virtual-drive/tree/infrastructure/CryptoJsNameDecrypt';
 import { getUser } from '../../../main/auth/service';
 import { SyncEngineIPC } from '../../SyncEngineIpc';
 import { ItemsContainer } from './ItemsContainer';
