+++ src/context/virtual-drive/tree/application/ItemsSearcher.ts
