+++ src/context/virtual-drive/tree/domain/FileNode.ts
