+++ src/context/virtual-drive/tree/domain/ItemsIndexedByPath.ts
